requirements.txt
# Required dependencies
teams-ai>=1.0.0
botbuilder-core>=4.16.0
botbuilder-schema>=4.16.0
aiohttp>=3.9.0
gunicorn>=21.2.0

# Azure OpenAI (Legacy client if needed, or via teams-ai)
# teams-ai usually handles this but we kept custom client for now
openai>=1.12.0

# File processing
pypdf>=4.0.0
python-docx>=1.1.0
httpx>=0.27.0

# Configuration
python-dotenv>=1.0.0
